crop-prediction/requirements.txt
fastapi 
uvicorn 
requests 
beautifulsoup4 
lxml
openai
python-dotenv
